pyproject.toml
[tool.poetry]
name = "myblog"
version = "0.1.0"
description = ""
authors = ["cbhcbhcbh <15862528607@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
django-cors-headers = "^4.2.0"
django-mdeditor = "^0.1.20"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pre-commit = "^3.4.0"
colorlog = "^6.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
